Building test backend

+++ backend/src/controllers/pictures.ts
@@ -4,9 +4,9 @@ import { tokenExtractor } from '../middleware/tokenExtractor.js';
 import { pictureFinder } from '../middleware/finders.js';
 import { validatePictureUploadInput } from '../middleware/validateInput.js';
 import { handlePictureUpdates } from '../middleware/validateUpdateInput.js';
-import { createThumbnail } from '../middleware/createThumbnail.js';
+import { writeFileCreateThumbnail } from '../middleware/createThumbnail.js';
+import { handleUpload } from '../middleware/uploadMiddleware.js';
 
-import { upload } from '../utils/multerConfig.js';
 import { picIncludeBasic, PicIncludeAll } from '../utils/includeOptions.js';
 import { pictureQueryOptions } from '../utils/queryHelpers.js';
 import { deleteFile } from '../utils/fileUtils.js';
@@ -31,9 +31,9 @@ const router = express.Router();
 // - validatePictureUploadInput: validates body input
 router.post(
   '/upload',
+  handleUpload,
   tokenExtractor,
-  upload.single('image'),
-  createThumbnail,
+  writeFileCreateThumbnail,
   validatePictureUploadInput,
   async (req: Request<object, object, PictureInput>, res: Response) => {
     const { type, textFi, textEn, year, month, keywords } = req.body;
@@ -171,7 +171,9 @@ router.delete(
     const thumbnailPath = getPath(picture.urlThumbnail || '');
 
     deleteFile(filePath);
-    deleteFile(thumbnailPath);
+    if (picture.urlThumbnail != null) {
+      deleteFile(thumbnailPath);
+    }
 
     await picture.destroy();
     res
